fix: silence -Wsign-conversion in pin_current_thread_to_cpu on gcc/Linux
CI's gcc/Release job (the only leg that actually compiles the
__linux__ branch of pin_current_thread_to_cpu -- macOS never does)
failed: CPU_SET(cpu_id, &cpu_set) implicitly converts cpu_id (int) to
size_t internally, which -Wsign-conversion correctly flags. Never
caught locally since this dev machine is macOS, where the function
takes the #else (no-op) branch entirely -- another instance of this
project's now-familiar pattern of gcc/Linux-only issues that only
CI can catch.

Fix: an explicit static_cast<std::size_t>(cpu_id) at the call site.
cpu_id is never expected to be negative in this codebase's usage (a
plain CPU index), so the cast is the correct, standard fix for this
warning class, not a workaround.

Verified: clean local rebuild in Release with benchmarks + -Werror
(the #else branch, all this machine can exercise); the actual fixed
line needs CI to confirm, matching this project's established pattern
for gcc/Linux-specific issues throughout this session.

+++ bench/latency_histogram.hpp
@@ -78,7 +78,7 @@ private:
 #ifdef __linux__
     cpu_set_t cpu_set;
     CPU_ZERO(&cpu_set);
-    CPU_SET(cpu_id, &cpu_set);
+    CPU_SET(static_cast<std::size_t>(cpu_id), &cpu_set);
     return pthread_setaffinity_np(pthread_self(), sizeof(cpu_set_t), &cpu_set) == 0;
 #else
     return false;
